+++ api/pmpanel/model.go
@@ -10,7 +10,7 @@ type NodeInfoResponse struct {
 	TrafficRate     float64 `json:"trafficRate"`
 	RawServerString string  `json:"outServer"`
 	Port            int     `json:"outPort"`
-	AlterId         int     `json:"alterId"`
+	AlterId         uint16  `json:"alterId"`
 	Network         string  `json:"network"`
 	Security        string  `json:"security"`
 	Host            string  `json:"host"`
